refactor(#187): extract assessment orchestration into AssessmentOrchestrator (parent #126)
Move the full happy-path orchestration out of Program.cs into a new
internal sealed class AssessmentOrchestrator (Orchestration/AssessmentOrchestrator.cs).
Program.Main now: parse args -> build config -> build DI -> create scope ->
resolve AssessmentOrchestrator -> RunAsync. Ctrl+C handling, outer try/catch,
BuildConfiguration, and ConfigureServices remain in Program for #188/#189.

Methods lifted (live code; 3 dead methods retained for #189 to delete):
DisplayWelcomeMessage, ValidateEffectiveConfiguration, RunAssessmentAsync,
GenerateOutputsAsync, SanitizeName, CombineCosmosAnalyses, CombineSqlAssessments,
CombineDataFactoryEstimates, DisplayCompletionMessage, ExtractAccountNameFromEndpoint,
GetUserInputsAsync, DiscoverAllDatabasesAsync, GetOutputDirectoryAsync,
AutoDiscoverMonitoringAsync, TestConnectionAsync, GenerateSqlProjectAsync.

POCOs UserInputs and MonitoringConfiguration moved to Models/UserInputs.cs
(CosmosToSqlAssessment.Models namespace).

AzureMonitorTestQuery / AzureMonitorTestQueryDays constants relocated to
the orchestrator (only TestConnectionAsync consumes them).

Program.ConfigureServices widened from private static to internal static so the
orchestrator's per-database override branch can rebuild services with a
swapped Cosmos endpoint. This temporary coupling will be eliminated in #188.

Behavior change: logger category for orchestration code is now
ILogger<AssessmentOrchestrator> (was ILogger<Program>). Anyone filtering
logs on the old category will need to update.

Pre-existing bugs preserved 1-for-1 (follow-up issues to be filed at parent
close):
  1. Per-database override config provider ordering reverses the endpoint
     override (.AddInMemoryCollection(...).AddConfiguration(configuration) -
     later provider wins).
  2. OperationCanceledException is swallowed by broad catch (Exception) blocks
     in the Cosmos/SQL/DataFactory phases of RunAssessmentAsync.
  3. Override ServiceProvider is disposed at the bottom of the loop without a
     try/finally, so mid-loop failures leak it.

Program.cs: 1340 -> 365 lines.
Build: clean. Tests: 297/297 pass (Wave-1 #125 E2E harness included).
